Cleanup integration_test, removing unnecessary mod
